+++ CHANGELOG.md
@@ -4,13 +4,13 @@
 
 ### Added
 
-- Added the `sprig.swapOob()` function that swaps a template out-of-band.
-- Added the `sprig.triggerRefresh()` function that triggers a refresh on a Sprig component.
+- Added the `sprig.swapOob(selector, template, variables)` function that swaps a template out-of-band.
+- Added the `sprig.triggerRefresh(selector, variables)` function that triggers a refresh on a Sprig component.
 
 ### Changed
 
 - Updated htmx to a pre-release version 2.0.2.
-- The `sprig.triggerRefreshOnLoad()` function now appends output to the end of the body and should be called using `{% do sprig.triggerRefreshOnLoad() %}`.
+- The `sprig.triggerRefreshOnLoad(selector)` function now appends output to the end of the body and should be called using `{% do sprig.triggerRefreshOnLoad(selector) %}`.
 
 ## 2.8.0 - 2024-07-15